Rework export tests to use node:test natively
